Move to standalone CMake.
